refactor(formatDate, formatTimeAgo): createdAt 관련 비즈니스 로직 분리
